chore: Rename module mqtt-data to mqtt-data-params
