+++ settings.gradle.kts
@@ -26,7 +26,7 @@ include(":spi:http-tls:http-tls-client-spi")
 include(":spi:http-tls:http-tls-data-address-spi")
 include(":spi:mqtt:mqtt-data-address-spi")
 include(":spi:mqtt:mqtt-data-endpoint-spi")
-include(":spi:mqtt:mqtt-data-spi")
+include(":spi:mqtt:mqtt-data-params-spi")
 
 // extensions - control plane
 include(":edc-extensions:fx-json-ld-core")
