+++ spi/mqtt/mqtt-data-params-spi/build.gradle.kts
