+++ spi/mqtt/mqtt-data-params-spi/src/main/java/org/factoryx/edc/mqtt/data/params/spi/MqttConstants.java
@@ -17,19 +17,19 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.factoryx.edc.mqtt.data.spi;
+package org.factoryx.edc.mqtt.data.params.spi;
 
-import org.factoryx.edc.mqtt.data.address.spi.MqttDataAddress;
+public interface MqttConstants {
 
-public interface MqttParamsProvider {
+    String DSP_MQTT_PREFIX = "dsp-mqtt";
+    String DSP_MQTT_NS = "https://w3id.org/dspace/2025/1/mqtt-pull/";
 
-    /**
-     * Register decorator
-     */
-    void registerSourceDecorator(MqttParamsDecorator decorator);
+    String USERNAME = DSP_MQTT_NS + "username";
+    String PASSWORD = DSP_MQTT_NS + "password";
 
-    /**
-     * Provide params for Mqtt Data Transfer Type
-     */
-    MqttParams provideParams(MqttDataAddress dataAddress);
+    String REFRESH_ENDPOINT = DSP_MQTT_NS + "refreshEndpoint";
+    String REFRESH_TOKEN = DSP_MQTT_NS + "refreshToken";
+    String EXPIRES_IN = DSP_MQTT_NS + "expiresIn";
+    String AUTHORIZATION = DSP_MQTT_NS + "authorization";
+    String REFRESH_AUDIENCE = DSP_MQTT_NS + "refreshAudience";
 }
